refactor: getannouncedgroupproducer and getannouncedgroupuser
